LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/light-dismiss-event-ordering-expected.txt

target

PASS Tests the interactions between popover light dismiss and pointer/mouse events. eventName: pointerdown, capture: true
FAIL Tests the interactions between popover light dismiss and pointer/mouse events. eventName: pointerup, capture: true assert_equals: The popover should be closed via light dismiss even when preventDefault is called. expected 0 but got 1
PASS Tests the interactions between popover light dismiss and pointer/mouse events. eventName: mousedown, capture: true
PASS Tests the interactions between popover light dismiss and pointer/mouse events. eventName: mouseup, capture: true
PASS Tests the interactions between popover light dismiss and pointer/mouse events. eventName: click, capture: true
PASS Tests the interactions between popover light dismiss and pointer/mouse events. eventName: pointerdown, capture: false
FAIL Tests the interactions between popover light dismiss and pointer/mouse events. eventName: pointerup, capture: false assert_equals: The popover should be closed via light dismiss even when preventDefault is called. expected 0 but got 1
PASS Tests the interactions between popover light dismiss and pointer/mouse events. eventName: mousedown, capture: false
PASS Tests the interactions between popover light dismiss and pointer/mouse events. eventName: mouseup, capture: false
PASS Tests the interactions between popover light dismiss and pointer/mouse events. eventName: click, capture: false
FAIL Tests the order of pointer/mouse events during popover light dismiss. assert_array_equals: pointer and popover events should be fired in the correct order. expected property 2 to be "beforetoggle newState: closed" but got "pointerup" (expected array ["pointerdown", "mousedown", "beforetoggle newState: closed", "pointerup", "mouseup", "click"] got ["pointerdown", "mousedown", "pointerup", "beforetoggle newState: closed", "mouseup", "click"])
